Move external links controllers and templates under observations
