+++ app/views/observations/show/_links.html.erb
@@ -1,5 +1,5 @@
-<%= tag.div(class: "panel panel-default",
-            id: "observation_links_panel") do
+<%=
+tag.div(class: "panel panel-default", id: "observation_links_panel") do
 
   [
     tag.div(class: "panel-heading") do
@@ -8,45 +8,41 @@
 
     tag.turbo_frame(
       id: "observation_external_links", class: "panel-body",
-      data: { controller: "section-update",
-              updated_by: "modal_external_link" }
+      data: { controller: "section-update", updated_by: "modal_external_link" }
     ) do
 
       if obs.external_links.any?
-        [
-          tag.ul(class: "tight-list") do
-            obs.external_links.sort_by(&:site_name).each do |link|
-              tag.li(id: "external_link_#{link.id}") do
-                concat(link_to(link.external_site.name, link.url))
-                if link.can_edit? || in_admin_mode?
-                  concat(
-                    [
-                      " [",
-                      modal_link_to("external_link",
-                                    *edit_external_link_tab(link: link)),
-                      "|",
-                      destroy_button(
-                        name: :destroy_object.t(type: :external_link),
-                        target: external_link_path(id: link.id),
-                        data: {
-                          turbo: true,
-                          confirm: :show_observation_remove_link_dialog.l
-                        }, icon: :remove,
-                        class: "destroy_external_link_link_#{link.id}"
-                      )
-                      "]"
-                    ].safe_join(" ")
-                  )
-                end
-              end
-            end.safe_join
-          end
-        ].safe_join
+        tag.ul(class: "tight-list") do
+          obs.external_links.sort_by(&:site_name).map do |link|
+            tag.li(id: "external_link_#{link.id}") do
+              concat(link_to(link.external_site.name, link.url))
+              concat(
+                [
+                  " [",
+                  modal_link_to("external_link",
+                                *edit_external_link_tab(link: link)),
+                  "|",
+                  destroy_button(
+                    name: :destroy_object.t(type: :external_link),
+                    target: external_link_path(id: link.id),
+                    icon: :remove,
+                    # data: {
+                    #   turbo: true,
+                    #   turbo_confirm: :show_observation_remove_link_dialog.l
+                    # },
+                    class: "destroy_external_link_link_#{link.id}"
+                  ),
+                  "]"
+                ].safe_join(" ")
+              ) if link.can_edit? || in_admin_mode?
+            end
+          end.safe_join
+        end
       end
     end,
 
     tag.div(class: "panel-footer") do
-      concat(@new_sites.sort_by(&:name).each do |site|
+      @new_sites.sort_by(&:name).map do |site|
         [
           tag.span(site.name),
           [ # any html helper has to be safe_joined separately
@@ -56,7 +52,7 @@
             "]"
           ].safe_join
         ].safe_join(" ")
-      end)
+      end.safe_join
     end
 
   ].safe_join
