+++ db/schema.rb
@@ -491,7 +491,7 @@ ActiveRecord::Schema.define(version: 2023_10_19_202655) do
     t.integer "project_id", null: false
   end
 
-  create_table "project_members", charset: "utf8mb4", collation: "utf8mb4_0900_ai_ci", force: :cascade do |t|
+  create_table "project_members", charset: "utf8mb3", force: :cascade do |t|
     t.integer "project_id"
     t.integer "user_id"
     t.boolean "trusted", default: false, null: false
